Fix conflict of types in gmtp_write_xmit

+++ linux-3.17.1/net/gmtp/proto.c
@@ -206,7 +206,7 @@ int gmtp_init_sock(struct sock *sk)
 	gp->tx_first_stamp	= 0UL;
 	gp->tx_last_stamp	= 0UL;
 	gp->tx_max_rate		= 0UL; /* Unlimited */
-	gp->tx_byte_budget	= LONG_MIN;
+	gp->tx_byte_budget	= INT_MIN;
 	gp->tx_adj_budget	= 0;
 
 	memset(gp->flowname, 0, GMTP_FLOWNAME_LEN);
